Safer shader

+++ modules/carto/src/layers/raster-layer-vertex.glsl.ts
@@ -32,8 +32,8 @@ void main(void) {
   vec2 common_position = offset.xy;
   float scale = offset.z;
 
-  int yIndex = - (gl_InstanceID / 256);
-  int xIndex = gl_InstanceID + (yIndex * 256);
+  float yIndex = - float(gl_InstanceID / 256);
+  float xIndex = float(gl_InstanceID) + (yIndex * 256.0);
   common_position += scale * vec2(xIndex, yIndex);
 
   vec4 color = isStroke ? instanceLineColors : instanceFillColors;
